fixed some warnings
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein_uncertainty/trunk@155 ec50deee-bab5-4ae7-ba08-ff0b0ff4da06

+++ src/UncertaintyDataCall.h
@@ -257,7 +257,7 @@ namespace megamol {
             }
             if (!this->secStructUncertainty)
                 return default;
-            else if (this->secStructUncertainty->Count() <= index)
+            else if (static_cast<int>(this->secStructUncertainty->Count()) <= index)
                 return default;
             else if (this->secStructUncertainty->operator[](index).Count() <= i)
                 return default;
@@ -280,7 +280,7 @@ namespace megamol {
             }
             if (!this->sortedSecStructAssignment)
                 return default;
-            else if (this->sortedSecStructAssignment->Count() <= index)
+            else if (static_cast<int>(this->sortedSecStructAssignment->Count()) <= index)
                 return default;
             else if (this->sortedSecStructAssignment->operator[](index).Count() <= i)
                 return default;
